Move and reuse StaticMethod helper tighten type support
